- Implement PROJECT_LOS and PROJECT_FLOOD. PROJECT_LOS is for attacks which can be used through line of sight transparent terrain in addition to line of fire transparent terrain. This allows these attacks to be fired through e.g. closed windows and is intended for attacks such as beams of light, and psychic powers where it should be possible to attack only if seeing the target.
PROJECT_FLOOD uses the standard Angband flooding algorithm to determine which grids can be affected in addition to any other flags. This is intended for e.g. gas clouds which spread out as opposed to explode out from the target grid. It can be combined with PROJECT_BEAM to create wide beams (At the moment as a hack).