- Implemented PROJECT_FORK and PROJECT_MYST which remove random parts of the projection.
For this process, we try to ensure that we have a minimum number of grids per range band to keep the overall shape of the projection intact.

PROJECT_FORK tries to ensure that no part of the projection	is disconnected and will remove empty grids ahead of grids which contain monsters. This should make lightning like shapes. PROJECT_MYST is doesn't have these restrictions.